Commit to Python3 only
Remove py23 and reimplement associated functions in Python 3.
