+++ src/cothread/load_ca.py
@@ -32,8 +32,6 @@
 # This file can also be run as a standalone script to discover the path to
 # libca.
 
-from __future__ import print_function
-
 import ctypes
 import platform
 import os
