+++ src/cothread/tools/pvtree.py
@@ -2,8 +2,6 @@
 
 # Simple tool for viewing the chain of PV dependencies.
 
-from __future__ import print_function
-
 import sys
 import re
 import os
